proper move detection
algorithms are implemented for every piece to decide where the pieces can move and where they can attack. the only thing that still needs to be added in terms of piece movement is not letting king move into a place that will put it in check, and not letting a player move a piece that would lead to a discovered check on the player.